+++ lib/item.ts
@@ -6,7 +6,7 @@ import * as MC from "@minecraft/server";
 export { ItemLockMode, ItemStack };
 export { ItemBag } from "./item/bag.js";
 export { ItemDurability } from "./item/durability.js";
-export { Enchantment, ItemEnchantments } from "./item/enchantments.js";
+export { EnchantmentType, ItemEnchantments } from "./item/enchantments.js";
 export { ItemTags } from "./item/tags.js";
 export { ItemType } from "./item/type.js";