+++ lib/item/enchantments.ts
@@ -1,10 +1,16 @@
 import { Wrapper } from "../wrapper.js";
 import { EnchantmentType } from "@minecraft/server";
+import * as I from "../inspect.js";
+import * as PP from "../pprint.js";
 import * as MC from "@minecraft/server";
 
 export { EnchantmentType };
 
-export class ItemEnchantments extends Wrapper<MC.ItemEnchantableComponent|undefined> implements Set<Enchantment> {
+export class ItemEnchantments extends Wrapper<MC.ItemEnchantableComponent|undefined> implements Map<EnchantmentType, number> {
+    public static "type"(enchantmentId: string): EnchantmentType|undefined {
+        return MC.EnchantmentTypes.get(enchantmentId);
+    }
+
     public get size(): number {
         // FIXME: Remove this glue code when the API is updated to 1.9.0.
         if (this.raw && "enchantments" in this.raw) {
@@ -24,36 +30,19 @@ export class ItemEnchantments extends Wrapper<MC.ItemEnchantableComponent|undefi
         return "ItemEnchantments";
     }
 
-    public [Symbol.iterator](): IterableIterator<Enchantment> {
-        return this.values();
+    public [Symbol.iterator](): IterableIterator<[EnchantmentType, number]> {
+        return this.entries();
     }
 
-    public add(ench: Enchantment): this {
-        if (this.raw) {
-            // FIXME: Remove this glue code when the API is updated to 1.9.0.
-            if ("enchantments" in this.raw) {
-                // @ts-ignore
-                this.raw.enchantments.addEnchantment(ench);
-                return this;
-            }
-
-            this.raw.addEnchantment(ench);
-            return this;
-        }
-        else {
-            throw new TypeError("this item can have no enchantments");
-        }
-    }
-
-    public canAdd(ench: Enchantment): boolean {
+    public canSet(type: EnchantmentType|string, level: number): boolean {
         // FIXME: Remove this glue code when the API is updated to 1.9.0.
         if (this.raw && "enchantments" in this.raw) {
             // @ts-ignore
-            return this.raw.enchantments.canAddEnchantment(ench);
+            return this.raw.enchantments.canAddEnchantment(new MC.Enchantment(type, level));
         }
 
         return this.raw
-            ? this.raw.canAddEnchantment(ench)
+            ? this.raw.canAddEnchantment({type, level})
             : false;
     }
 
@@ -63,23 +52,16 @@ export class ItemEnchantments extends Wrapper<MC.ItemEnchantableComponent|undefi
         }
     }
 
-    public delete(ench: Enchantment|EnchantmentType|string): boolean {
-        if (this.has(ench)) {
+    public delete(type: EnchantmentType|string): boolean {
+        if (this.has(type)) {
             // FIXME: Remove this glue code when the API is updated to 1.9.0.
             if (this.raw && "enchantments" in this.raw) {
-                if (ench instanceof Enchantment)
-                    // @ts-ignore
-                    this.raw.enchantments.removeEnchantment(ench.type);
-                else
-                    // @ts-ignore
-                    this.raw.enchantments.removeEnchantment(ench);
+                // @ts-ignore
+                this.raw.enchantments.removeEnchantment(type);
                 return true;
             }
 
-            if (ench instanceof Enchantment)
-                this.raw!.removeEnchantment(ench.type);
-            else
-                this.raw!.removeEnchantment(ench);
+            this.raw!.removeEnchantment(type);
             return true;
         }
         else {
@@ -87,122 +69,114 @@ export class ItemEnchantments extends Wrapper<MC.ItemEnchantableComponent|undefi
         }
     }
 
-    public *entries(): IterableIterator<[Enchantment, Enchantment]> {
-        for (const ench of this) {
-            yield [ench, ench];
+    public *entries(): IterableIterator<[EnchantmentType, number]> {
+        if (this.raw) {
+            // FIXME: Remove this glue code when the API is updated to 1.9.0.
+            if (this.raw && "enchantments" in this.raw) {
+                // @ts-ignore
+                for (const ench of this.raw.enchantments) {
+                    yield [ench.type, ench.level];
+                }
+                return;
+            }
+
+            for (const ench of this.raw.getEnchantments()) {
+                yield [ench.type as EnchantmentType, ench.level];
+            }
         }
     }
 
-    public forEach(f: (value: Enchantment, value2: Enchantment, set: Set<Enchantment>) => void, thisArg?: any): void {
+    public forEach(f: (value: number, key: EnchantmentType, map: ItemEnchantments) => void, thisArg?: any): void {
         const boundF = f.bind(thisArg);
-        for (const ench of this) {
-            boundF(ench, ench, this);
+        for (const [type, level] of this) {
+            boundF(level, type, this);
         }
     }
 
-    public get(type: EnchantmentType|string): Enchantment|undefined {
+    public "get"(type: EnchantmentType|string): number|undefined {
         if (this.raw) {
             // FIXME: Remove this glue code when the API is updated to 1.9.0.
             if ("enchantments" in this.raw) {
                 // @ts-ignore
-                const raw = this.raw.enchantments.getEnchantment(type);
-                return raw ? new Enchantment(raw) : undefined;
+                const ench = this.raw.enchantments.getEnchantment(type);
+                return ench ? ench.level : undefined;
             }
 
-            const raw = this.raw.getEnchantment(type);
-            return raw ? new Enchantment(raw) : undefined;
+            const ench = this.raw.getEnchantment(type);
+            return ench ? ench.level : undefined;
         }
         else {
             return undefined;
         }
     }
 
-    public has(ench: Enchantment|EnchantmentType|string): boolean {
+    public has(type: EnchantmentType|string): boolean {
         if (this.raw) {
-            if (ench instanceof Enchantment) {
-                const existing = this.get(ench.type);
-                return existing
-                    ? existing.level == ench.level
-                    : false;
+            // FIXME: Remove this glue code when the API is updated to 1.9.0.
+            if ("enchantments" in this.raw) {
+                // @ts-ignore
+                return this.raw.enchantments.hasEnchantment(type) > 0;
             }
-            else {
-                // FIXME: Remove this glue code when the API is updated to 1.9.0.
-                if ("enchantments" in this.raw) {
-                    // @ts-ignore
-                    return this.raw.enchantments.hasEnchantment(ench);
-                }
 
-                return this.raw.hasEnchantment(ench);
-            }
+            return this.raw.hasEnchantment(type);
         }
         else {
             return false;
         }
     }
 
-    public keys(): IterableIterator<Enchantment> {
-        return this.values();
-    }
-
-    public *values(): IterableIterator<Enchantment> {
+    public *keys(): IterableIterator<EnchantmentType> {
         if (this.raw) {
-            for (const raw of this.raw.getEnchantments()) {
-                yield new Enchantment(raw);
+            // FIXME: Remove this glue code when the API is updated to 1.9.0.
+            if (this.raw && "enchantments" in this.raw) {
+                // @ts-ignore
+                for (const ench of this.raw.enchantments) {
+                    yield ench.type;
+                }
+                return;
+            }
+
+            for (const ench of this.raw.getEnchantments()) {
+                yield ench.type as EnchantmentType;
             }
         }
     }
-}
 
-export class Enchantment implements MC.Enchantment {
-    #type: EnchantmentType;
-    #level: number;
-
-    /** Package private: user code should not use this. */
-    public constructor(rawEnchantment: MC.Enchantment);
-
-    /** Construct an enchantment. */
-    public constructor(type: EnchantmentType|string, level: number);
-
-    public constructor(...args: any[]) {
-        if (args[0] instanceof EnchantmentType) {
-            this.#type  = args[0];
-            this.#level = args[1];
-        }
-        else if (typeof args[0] === "string") {
-            const type = MC.EnchantmentTypes.get(args[0]);
-            if (type) {
-                this.#type  = type;
-                this.#level = args[1];
+    public "set"(type: EnchantmentType|string, level: number): this {
+        if (this.raw) {
+            // FIXME: Remove this glue code when the API is updated to 1.9.0.
+            if ("enchantments" in this.raw) {
+                // @ts-ignore
+                this.raw.enchantments.addEnchantment(new MC.Enchantment(type, level));
+                return this;
             }
-            else {
-                throw new Error(`Non-existent enchantment type: ${args[0]}`);
-            }
+
+            this.raw.addEnchantment({type, level});
+            return this;
         }
         else {
-            // MC.Enchantment is just an interface rather than a class,
-            // which complicates things here.
-            if (typeof args[0].type === "string") {
-                const type = MC.EnchantmentTypes.get(args[0].type);
-                if (type) {
-                    this.#type  = type;
-                    this.#level = args[0].level;
+            throw new TypeError("this item can have no enchantments");
+        }
+    }
+
+    public *values(): IterableIterator<number> {
+        if (this.raw) {
+            // FIXME: Remove this glue code when the API is updated to 1.9.0.
+            if (this.raw && "enchantments" in this.raw) {
+                // @ts-ignore
+                for (const ench of this.raw.enchantments) {
+                    yield ench.level;
                 }
-                else {
-                    throw new Error(`Non-existent enchantment type: ${args[0].type}`);
-                }
+                return;
             }
-            else {
-                this.#type  = args[0].type;
-                this.#level = args[0].level;
+
+            for (const ench of this.raw.getEnchantments()) {
+                yield ench.level;
             }
         }
     }
 
-    get type(): EnchantmentType {
-        return this.#type;
-    }
-
-    get level(): number {
-        return this.#level;
+    public [I.customInspectSymbol](inspect: (value: any, opts?: I.InspectOptions) => PP.Doc): PP.Doc {
+        return PP.spaceCat(PP.text("[ItemEnchantments]"), inspect(new Map(this)));
     }
 }